add new parameter "DatacenterId" into New-NetworkDomain, put "Location" as Alias attribute.
Removed DeployCaasNetworkDomainCmdlet.cs
Added NewCaasNetworkDomainCmdlet.cs
