+++ PowershellModule/CloudCompute.Powershell.Tests/CloudCompute.Powershell.Tests.csproj
@@ -100,6 +100,9 @@
     <None Include="RunTests.ps1">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
+    <None Include="ServerTests\NetworkDomain.Tests.ps1">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
     <None Include="ServerTests\ServerActions.Tests.ps1">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
